+++ paddle/fluid/pybind/eager_utils.cc
@@ -3102,7 +3102,7 @@ paddle::optional<Tensor*> GetInputOutTensorFromKwargs(PyObject* kwargs) {
     return paddle::none;
   }
   PyObject* obj = PyDict_GetItemString(kwargs, "out");
-  if (obj) {
+  if (obj && PyObject_TypeCheck(obj, p_tensor_type)) {
     return paddle::make_optional<paddle::Tensor*>(
         &(reinterpret_cast<TensorObject*>(obj)->tensor));
   }
